Pridané odstranenie profilu, ošetrenie hesiel v registračnom formulári pomocou js, pridané vyhľadávanie recpetov a predpriprava na zobrazenie konkrétneho receptu

+++ App/Views/recipes/findRecipe.view.php
@@ -20,13 +20,13 @@ if (count($data['recipes']) > 0){
                         <span class="badge bg-dark"><?=$recipe->getDuration()?></span>
                         <span class="badge bg-dark"><?=$recipe->getPortions()?> porcii</span>
 
-                        <p><?=substr($recipe->getProcess(),0,255)?>...</p>
+                        <p><?=substr($recipe->getProcess(),0,254)?>...</p>
 
                         <hr/>
-                        <a class="btn" href="?c=account&a=settingsForm">Otvoriť</a>
+                        <a class="btn" href="?c=recipes&a=showRecipe&id=<?=$recipe->getId()?>">Otvoriť</a>
 
-<!--                        Ak je prihlaseny uzivatel, moze upravit alebo odstranit recept-->
-                        <?php if($_SESSION['name'] != ''){ ?>
+<!--                    Ak je prihlaseny uzivatel, moze upravit alebo odstranit recept-->
+                        <?php if(isset($_SESSION['name']) && $_SESSION['name'] != ''){ ?>
                             <a class="btn btn-success" href="?c=account&a=showRecipe">Upraviť</a>
                             <a class="btn btn-danger" href="c=account&a=deleteRecipe">Odstrániť</a>
                         <?php }?>
@@ -45,13 +45,17 @@ if (count($data['recipes']) > 0){
     <div class="row">
         <div class="col-sm-2"></div>
 
-        <div class="col-sm-8">
+        <div class="col-sm-8 block">
             <div class="alert alert-danger" role="alert">
-                Zatiaľ si nepridal žiaden recept...
-                <a class="btn float-right" href="#">
-                    Pridať recept
-                </a>
+                Pre zadaný výraz sa nenašiel žiaden recept :(
             </div>
+
+            <form method="post" class="input-group mb-3" action="?c=recipes&a=findRecipe">
+                <input type="text" class="form-control" name="title"
+                       placeholder="Vyhľadať na základe názvu receptu...">
+                <button class="btn" type="submit">Vyhľadaj</button>
+            </form>
+
         </div>
 
         <div class="col-sm-2"></div>
